fraudlens/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
python-multipart==0.0.6
pydantic==2.5.3
Pillow==10.2.0
httpx==0.26.0
numpy==1.26.3
scipy==1.11.4
piexif==1.1.3
PyExifTool==0.5.6
google-cloud-vision==3.7.2
c2pa-python==0.4.0
loguru==0.7.2
mediapipe==0.10.9
opencv-python-headless==4.9.0.80
fpdf2==2.7.9
pytesseract==0.3.10
